moved allNuggets creation into helper
